requirements.txt
customtkinter>=5.2.2
matplotlib>=3.10.8
numpy>=2.4.2
scipy>=1.17.0
trimesh>=4.11.1
ezdxf>=1.4.3
svgwrite>=1.4.3
libigl>=2.6.1
networkx>=3.6
shapely>=2.0
PySide6>=6.7
qt-material>=2.14
gmsh>=4.13
pyqtgraph>=0.14.0
PyOpenGL>=3.1.7
